+++ src/index.ts
@@ -1,15 +1,23 @@
-import { unlink, write, eventEmitter, Options, Callback} from './lib/secure-rm'
+import { unlink, Options, Callback} from './lib/secure-rm'
+import { validIDs, standards, UnlinkStandard } from './lib/standards'
+import { write, eventEmitter } from './lib/write'
 
 interface SecureRm {
   (path: string, options?: Options | Callback, callback?: Callback): typeof unlink;
   event: typeof eventEmitter;
   write: typeof write;
-  default: SecureRm
+  standards: typeof standards;
+  validIDs: typeof validIDs;
+  UnlinkStandard: typeof UnlinkStandard;
+  default: SecureRm;
 }
 
 let secureRmExport = <unknown>unlink as SecureRm
 secureRmExport.event = eventEmitter
 secureRmExport.write = write
+secureRmExport.standards = standards
+secureRmExport.validIDs = validIDs
+secureRmExport.UnlinkStandard = UnlinkStandard
 secureRmExport.default = secureRmExport
 
 export = secureRmExport