+++ src/lib/secure-rm.ts
@@ -1,18 +1,17 @@
 import fs from 'fs'
 import rimraf from 'rimraf'
-import { methods, validIDs } from './methods'
-export { write, eventEmitter } from './write'
+import { standards, validIDs } from './standards'
 
 export interface Options {
-  method?: string
-  customMethod?: typeof fs.unlink
+  standard?: string
+  customStandard?: typeof fs.unlink
   maxBusyTries?: number
   disableGlob?: boolean
 }
 
 export interface ParsedOptions {
-  method: keyof typeof methods
-  customMethod?: typeof fs.unlink
+  standard: keyof typeof standards
+  customStandard?: typeof fs.unlink
   maxBusyTries?: number
   disableGlob?: boolean
 }
@@ -29,10 +28,10 @@ export function unlink(path: string, options?: Options | Callback, callback?: Ca
   // Parse if callback is provided
   if (callback === undefined && typeof options === 'function') {
     callback = options
-    options = { method: 'secure' }
+    options = { standard: 'secure' }
   }
-  // Define method if none is provided
-  if (typeof (options as Options).method === undefined) (options as Options).method = 'secure'
+  // Define standard if none is provided
+  if (typeof (options as Options).standard === undefined) (options as Options).standard = 'secure'
 
   if (callback) unlinkCallback(path, options as ParsedOptions, (err: NodeJS.ErrnoException | null, path: string) => callback!(err, path))
   else return unlinkPromise(path, options as ParsedOptions)
@@ -42,38 +41,38 @@ export function unlink(path: string, options?: Options | Callback, callback?: Ca
 
 // Callback version
 function unlinkCallback(path: string, options: ParsedOptions, callback: Callback): void {
-  if (options.customMethod) {
+  if (options.customStandard) {
     rimraf(path, {
-      unlink: options.customMethod,
+      unlink: options.customStandard,
       maxBusyTries: options.maxBusyTries || 3,
       disableGlob: options.disableGlob || false
     }, (err: NodeJS.ErrnoException) => callback(err, path))
-  } else if (validIDs.includes(options.method)) {
+  } else if (validIDs.includes(options.standard)) {
     rimraf(path, {
-      unlink: methods[options.method].method,
+      unlink: standards[options.standard].method,
       maxBusyTries: options.maxBusyTries || 3,
       disableGlob: options.disableGlob || false
     }, (err: NodeJS.ErrnoException) => callback(err, path))
   } else {
-    callback(new Error(`'${options.method}' is not a valid ID. \nList of valid IDs: ${validIDs}`), path)
+    callback(new Error(`'${options.standard}' is not a valid ID. \nList of valid IDs: ${validIDs}`), path)
   }
 }
 
 // Promise version
 function unlinkPromise(path: string, options: ParsedOptions): Promise<string> {
   return new Promise((resolve, reject) => {
-    if (options.customMethod) {
+    if (options.customStandard) {
       rimraf(path, {
-        unlink: options.customMethod,
+        unlink: options.customStandard,
         maxBusyTries: options.maxBusyTries || 3,
         disableGlob: options.disableGlob || false
       }, (err: NodeJS.ErrnoException) => {
         if (err) reject(err)
         else resolve(path)
       })
-    } else if (validIDs.includes(options.method)) {
+    } else if (validIDs.includes(options.standard)) {
       rimraf(path, {
-        unlink: methods[options.method].method,
+        unlink: standards[options.standard].method,
         maxBusyTries: options.maxBusyTries || 3,
         disableGlob: options.disableGlob || false
       }, (err: NodeJS.ErrnoException) => {
@@ -81,7 +80,7 @@ function unlinkPromise(path: string, options: ParsedOptions): Promise<string> {
         else resolve(path)
       })
     } else {
-      reject(new Error(`'${options.method}' is not a valid ID. \nList of valid IDs: ${validIDs}`))
+      reject(new Error(`'${options.standard}' is not a valid ID. \nList of valid IDs: ${validIDs}`))
     }
   })
 }