+++ src/lib/write.ts
@@ -41,7 +41,7 @@ const write = {
       const buffer = Buffer.alloc(fileSize)
       crypto.randomFill(buffer, (err, randomBuffer) => {
         if (err) reject(err)
-        eventEmitter.emit('info', file, 'Writing random data ')
+        eventEmitter.emit('verbose', file, 'Writing random data ')
         fs.writeFile(file, randomBuffer, (err) => {
           if (err) reject(err)
           else if (passes > 1) {
@@ -55,7 +55,7 @@ const write = {
   },
   zeroes: (file: string, fileSize: number, passes: number = 1): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, 'Writing zeroes ')
+      eventEmitter.emit('verbose', file, 'Writing zeroes ')
       fs.writeFile(file, Buffer.alloc(fileSize, 0b00000000), (err) => {
         if (err) reject(err)
         else if (passes > 1) {
@@ -68,7 +68,7 @@ const write = {
   },
   ones: (file: string, fileSize: number, passes: number = 1): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, 'Writing ones ')
+      eventEmitter.emit('verbose', file, 'Writing ones ')
       fs.writeFile(file, Buffer.alloc(fileSize, 0b11111111), (err) => {
         if (err) reject(err)
         else if (passes > 1) {
@@ -82,7 +82,7 @@ const write = {
   // Write one byte on the whole file
   byte: (file: string, data: number, fileSize: number, passes: number = 1): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, `Writing 0x${data.toString(16)} `)
+      eventEmitter.emit('verbose', file, `Writing 0x${data.toString(16)} `)
       fs.writeFile(file, Buffer.alloc(fileSize, data), (err) => {
         if (err) reject(err)
         else if (passes > 1) {
@@ -96,7 +96,7 @@ const write = {
   // Write an array of bytes on the whole file
   bytes: (file: string, dataArray: number[], fileSize: number, passes: number = 1): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, `Writing ${dataArray.map(x => '0x' + x.toString(16))} `)
+      eventEmitter.emit('verbose', file, `Writing ${dataArray.map(x => '0x' + x.toString(16))} `)
       // const dataConverted = Uint8Array.from(dataArray)
       const dataConverted = Buffer.from(dataArray)
       fs.writeFile(file, Buffer.alloc(fileSize, dataConverted), (err) => {
@@ -113,7 +113,7 @@ const write = {
   // And then cycle through the array
   cycleBytes: (file: string, dataArray: number[], fileSize: number, passes: number = 1): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, `Writing ${dataArray.map(x => '0x' + x.toString(16))} `)
+      eventEmitter.emit('verbose', file, `Writing ${dataArray.map(x => '0x' + x.toString(16))} `)
       // const dataConverted = Uint8Array.from(dataArray)
       const dataConverted = Buffer.from(dataArray)
       fs.writeFile(file, Buffer.alloc(fileSize, dataConverted), (err) => {
@@ -130,7 +130,7 @@ const write = {
   // And then + increment until end byte
   incrementByte: (file: string, { start, end, increment }: ForInterface, fileSize: number): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, `Writing 0x${start.toString(16)} `)
+      eventEmitter.emit('verbose', file, `Writing 0x${start.toString(16)} `)
       fs.writeFile(file, Buffer.alloc(fileSize, start), (err) => {
         if (err) reject(err)
         else if (start + increment <= end) {
@@ -143,7 +143,7 @@ const write = {
   },
   randomByte: (file: string, fileSize: number, passes: number = 1): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, 'Writing random byte ')
+      eventEmitter.emit('verbose', file, 'Writing random byte ')
       fs.writeFile(file, Buffer.alloc(fileSize, crypto.randomBytes(1)[0]), (err) => {
         if (err) reject(err)
         else if (passes > 1) {
@@ -157,15 +157,15 @@ const write = {
   // Write the binary complement
   complementary: (file: string, fileSize: number): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, 'Reading file ')
+      eventEmitter.emit('verbose', file, 'Reading file ')
       fs.readFile(file, (err, data) => {
         if (err) reject(err)
         else {
-          eventEmitter.emit('info', file, 'Getting binary complement ')
+          eventEmitter.emit('verbose', file, 'Getting binary complement ')
           for (let i = 0, l = fileSize; i < l; i++) {
             data[i] = ~data[i]
           }
-          eventEmitter.emit('info', file, 'Writing binary complement ')
+          eventEmitter.emit('verbose', file, 'Writing binary complement ')
           fs.writeFile(file, data, (err) => {
             if (err) reject(err)
             else resolve({ file, fileSize })
@@ -179,7 +179,7 @@ const write = {
     return new Promise((resolve, reject) => {
       const newName = crypto.randomBytes(9).toString('base64').replace(/\//g, '0').replace(/\+/g, 'a')
       const newPath = path.join(path.dirname(file), newName)
-      eventEmitter.emit('info', file, `Renaming to ${newName} `)
+      eventEmitter.emit('verbose', file, `Renaming to ${newName} `)
       fs.rename(file, newPath, (err) => {
         if (err) reject(err)
         else resolve({ file: newPath, fileSize })
@@ -189,7 +189,7 @@ const write = {
   // Truncate to between 25% and 75% of the file size
   truncate: (file: string, fileSize: number): Promise<FileInfo> => {
     return new Promise((resolve, reject) => {
-      eventEmitter.emit('info', file, 'Truncating ')
+      eventEmitter.emit('verbose', file, 'Truncating ')
       const newSize = Math.floor((Math.random() * 0.5 + 0.25) * fileSize)
       fs.truncate(file, newSize, (err) => {
         if (err) reject(err)